test: move number test code
