+++ src/test/java/edu/lotto/automatic/controller/AutomaticLottoTest.java
@@ -14,15 +14,4 @@ import java.util.regex.Pattern;
  */
 public class AutomaticLottoTest {
 
-	@ParameterizedTest
-	@CsvSource(value = {"1:true", "1000:true", "100won:false", "qwerty:false"}, delimiter = ':')
-	void isNumber(String value, boolean expected) {
-		assertThat(Pattern.matches("[0-9]+$", value)).isEqualTo(expected);
-	}
-
-	@ParameterizedTest
-	@CsvSource(value = {"1:false", "100:false", "999:false", "1000:true", "9999:true", "10000:true"}, delimiter = ':')
-	void isMoreThanThousand(int value, boolean expected) {
-		assertThat(value >= 1000).isEqualTo(expected);
-	}
 }
